agent/requirements.txt
openai>=1.82.0
mcp>=1.9.0
python-dotenv>=1.0.0
